ai-service/requirements.txt
flask
flask-cors
requests
flask-limiter
python-dotenv
